+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Yii Framework 2 authclient extension Change Log
 ------------------------
 
 - Bug #330: OpenID Connect client now defaults to `'client_secret_basic'` in case `token_endpoint_auth_methods_supported` isn't specified (rhertogh)
-
+- Bug #332: OpenID Connect `aud` claim can either be a string or a list of strings (azmeuk)
 
 2.2.11 August 09, 2021
 ----------------------