+++ CHANGELOG.md
@@ -4,8 +4,9 @@ Yii Framework 2 authclient extension Change Log
 2.2.12 under development
 ------------------------
 
-- no changes in this release.
-
+- Bug #330: OpenID Connect client now defaults to `'client_secret_basic'` in case `token_endpoint_auth_methods_supported` isn't specified (rhertogh)
+- Bug #331: OpenID Connect `aud` claim can either be a string or a list of strings (azmeuk)
+- Bug #332: OpenID Connect `aud` nonce is passed from the authentication request to the token request (azmeuk)
 
 2.2.11 August 09, 2021
 ----------------------